Moved all code and almost all constants related to computing and drawing spectrum views to ui_spectrum. Did a few optimizations on the Spectrum Scope code, should run much faster in most cases. Enabling the optimization for the ui_lcd_hy28.c gave at least the SPI display a real performance boost.
